# Conflicts:
#	Client/Include/Client.vcxproj

+++ Client/Code/Shadow_Item.cpp
@@ -39,15 +39,18 @@ _int CShadow_Item::Update_Object(const _float& fTimeDelta)
 {
 	_int iExit = __super::Update_Object(fTimeDelta);
 
-	if (!m_pOwnerobject->Is_Active())
+	if (m_pOwnerobject->Is_Active())
 	{
-		CEventMgr::GetInstance()->Delete_Obj(this);
-		return iExit;
+		if (!m_pOwnerobject->Is_Active())
+		{
+			CEventMgr::GetInstance()->Delete_Obj(this);
+			return iExit;
+		}
+
+		Engine::Add_RenderGroup(RENDER_ALPHA, this);
+
+		m_fSize = fabs(m_InitY - m_pOwnerobject->Get_Transform()->Get_Info(INFO_POS).y);
 	}
-
-	Engine::Add_RenderGroup(RENDER_ALPHA, this);
-
-	m_fSize = fabs(m_InitY - m_pOwnerobject->Get_Transform()->Get_Info(INFO_POS).y);
 
 	return iExit;
 }
